Revert "Merge branch 'simplify_service_which_baseon_block_log' into dev"
This reverts commit f55c6ae03e1a26bc14c842c22acdc5c0ef7634bc, reversing
changes made to 578f281ad00e444392b13f82b4c7cc8c6458b89a.
