--- iservices/iotrxserver.go
@@ -1,3 +1,0 @@
-package iservices
-
-const IOTrxServiceName = "iotrx_svc"
